app/pyproject.toml
[project]
name = "explorer"
version = "0.1.0"
description = "Elefant Explorer"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.2",
    "django-allauth>=65.9.0",
    "django-environ>=0.12.0",
    "django-extensions>=4.1.0",
    "django-vite>=3.1.0",
    "mysqlclient>=2.2.7",
]

[dependency-groups]
dev = [
    "bpython>=0.25",
    "django-debug-toolbar>=5.0.1",
    "djlint>=1.36.4",
]
prod = [
    "gunicorn>=23.0.0",
    "sentry-sdk[django]>=2.23.1",
]
